+++ setup-env.sh
@@ -4,7 +4,7 @@ echo "$(date): Starting setup-env.sh"
 
 export CDK_VERSION=1.83.0
 export PACKER_VERSION=1.5.5
-export TASKCAT_VERSION=0.9.18
+export TASKCAT_VERSION=0.9.21
 
 # system upgrades and tools
 export DEBIAN_FRONTEND=noninteractive