+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 * Remove ElastiCache
 * Upgrade CDK to 1.83.0
 * Require ACM Certificate
+* Taskcat testing setup